requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
google-api-client
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2==0.18.1
Jinja2==2.11.2
oauthlib==3.1.0
PyVirtualDisplay==1.3.2
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
psycopg2-binary
fake-useragent